Fix `resolve()` method ambiguity
The ``$parameters` argument wasn't part of interface, but was implemented.
It introduced some ambiguity to method behaviour. It would behave
differently depending whether the class instance was bound to the container
before or not. Which means the calling code will be uncertain on the
`$parameters` array being used or not.

The only way to fix it -- is to remove the argument.